Improved Dockerfile to get a small Docker image
Using Docker multi-stage builds and "builder pattern" to have a much smaller Docker image.
Original image was ~750MB, new one ~25MB. This is achieved by having a "build" stage which
uses the `golang:1.9` image to build the `git-sync` binary.

This binary is then copied into the stage used when running the docker image.
This stage is based on a very small `alpine/git` base which is just alpine with the `git`
binary.

See: https://docs.docker.com/develop/develop-images/multistage-build/